CHANGE HTML, CSS AND JS FILES
I have made various structural changes to the HTML and CSS files to change the calculator's design. Furthermore, I have added a function of getOperator() in the JS file to get the operator selected. I have also added a function to clear the display of the calculator. I have further declared a variable to track the steps in the calculation process. The calculator is currently functional, but it is yet to do calculations involving more than two numbers.