I HAVE STARTED OVER FROM THE BEGINNING
I decided to start over with the Javascript. I have also made slight changes to the HTML File to add a button for deletion and decimal. At this stage, the functions for displaying numbers, clearing the display and function for inclusion of a decimal are functional. The Function to display the operators is not working.